+++ src/RcppExports.cpp
@@ -140,8 +140,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // optimize_dkp_theta_rcpp
-Rcpp::List optimize_dkp_theta_rcpp(const arma::mat& Xnorm, const arma::mat& Y, const std::string& prior, const double r0, const arma::vec& p0, const std::string& loss, const std::string& kernel, const bool isotropic, const arma::mat& init_gamma, const arma::vec& lower, const arma::vec& upper, const int max_iter, const int n_threads);
-RcppExport SEXP _BKP_optimize_dkp_theta_rcpp(SEXP XnormSEXP, SEXP YSEXP, SEXP priorSEXP, SEXP r0SEXP, SEXP p0SEXP, SEXP lossSEXP, SEXP kernelSEXP, SEXP isotropicSEXP, SEXP init_gammaSEXP, SEXP lowerSEXP, SEXP upperSEXP, SEXP max_iterSEXP, SEXP n_threadsSEXP) {
+Rcpp::List optimize_dkp_theta_rcpp(const arma::mat& Xnorm, const arma::mat& Y, const std::string& prior, const double r0, const arma::vec& p0, const std::string& loss, const std::string& kernel, const bool isotropic, const arma::mat& init_gamma, const arma::vec& lower, const arma::vec& upper, const int max_iter, const int n_threads, const std::string& ess, Nullable<NumericVector> m_shepard_loo);
+RcppExport SEXP _BKP_optimize_dkp_theta_rcpp(SEXP XnormSEXP, SEXP YSEXP, SEXP priorSEXP, SEXP r0SEXP, SEXP p0SEXP, SEXP lossSEXP, SEXP kernelSEXP, SEXP isotropicSEXP, SEXP init_gammaSEXP, SEXP lowerSEXP, SEXP upperSEXP, SEXP max_iterSEXP, SEXP n_threadsSEXP, SEXP essSEXP, SEXP m_shepard_looSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -158,7 +158,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const arma::vec& >::type upper(upperSEXP);
     Rcpp::traits::input_parameter< const int >::type max_iter(max_iterSEXP);
     Rcpp::traits::input_parameter< const int >::type n_threads(n_threadsSEXP);
-    rcpp_result_gen = Rcpp::wrap(optimize_dkp_theta_rcpp(Xnorm, Y, prior, r0, p0, loss, kernel, isotropic, init_gamma, lower, upper, max_iter, n_threads));
+    Rcpp::traits::input_parameter< const std::string& >::type ess(essSEXP);
+    Rcpp::traits::input_parameter< Nullable<NumericVector> >::type m_shepard_loo(m_shepard_looSEXP);
+    rcpp_result_gen = Rcpp::wrap(optimize_dkp_theta_rcpp(Xnorm, Y, prior, r0, p0, loss, kernel, isotropic, init_gamma, lower, upper, max_iter, n_threads, ess, m_shepard_loo));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -214,7 +216,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_BKP_optimize_lambda_bkp_rcpp", (DL_FUNC) &_BKP_optimize_lambda_bkp_rcpp, 9},
     {"_BKP_optimize_lambda_dkp_rcpp", (DL_FUNC) &_BKP_optimize_lambda_dkp_rcpp, 7},
     {"_BKP_optimize_bkp_theta_rcpp", (DL_FUNC) &_BKP_optimize_bkp_theta_rcpp, 16},
-    {"_BKP_optimize_dkp_theta_rcpp", (DL_FUNC) &_BKP_optimize_dkp_theta_rcpp, 13},
+    {"_BKP_optimize_dkp_theta_rcpp", (DL_FUNC) &_BKP_optimize_dkp_theta_rcpp, 15},
     {"_BKP_qbetabinom_rcpp", (DL_FUNC) &_BKP_qbetabinom_rcpp, 4},
     {"_BKP_get_twin_indices_rcpp", (DL_FUNC) &_BKP_get_twin_indices_rcpp, 5},
     {"_BKP_wendland_kernel_rcpp", (DL_FUNC) &_BKP_wendland_kernel_rcpp, 4},